apdet fungsi detail activity
- penambahan WeekAdapter.kt agar sebagai recycle view cekbok  list mingguan nya

-memperbaiki bug pada detailActivity yang sebelumnya hanya menampilkan nama barang saja, yg lain gak nampil

-memperbaiki intent pada MainActivity yang sebelumnya total waktu ber tipe double ke int agar tidak terjadi bug seperti sebelum nya

+++ app/src/main/java/com/example/projeklabmobile/WishlistItem.kt
@@ -4,6 +4,6 @@ data class WishlistItem(
     val id: String = "",
     val namaBarang: String = "",
     val hargaBarang: Double = 0.0,
-    val targetWaktu: Double = 0.0,
+    val targetWaktu: Int = 0,
     val progress: Int = 0 // buat persenan progresnya ygy
 )
